moved tech records repository
